backend/requirements.txt
Flask>=3.0.0
Flask-Login>=0.6.3
Flask-CORS>=4.0.0
Werkzeug>=3.0.1
pdf2image>=1.16.0
Pillow>=9.0.0
numpy>=1.21.0
phe>=1.5.0
pycryptodome>=3.15.0
python-docx>=0.8.11
pytest>=7.0.0
requests>=2.28.0